storagenode: fix disk monitoring for FreeBSD
